refactor: schedule chart
